Added docs to everything
