+++ util.go
@@ -1,6 +1,9 @@
 package goevo
 
-import "math/rand"
+import (
+	"math"
+	"math/rand"
+)
 
 func randomMapKey[T comparable, U any](m map[T]U) T {
 	n := rand.Intn(len(m))
@@ -36,4 +39,12 @@ func clamp(x, min, max float64) float64 {
 		return max
 	}
 	return x
+}
+
+func stdN(std float64) int {
+	v := math.Abs(rand.NormFloat64() * std)
+	if v > std*10 {
+		v = std * 10 // Lets just cap this at 10 std to prevent any sillyness
+	}
+	return int(math.Round(v))
 }
